Remove calculated variables from `input/income.py` Fixes #1828
