+++ changelog_entry.yaml
@@ -1,0 +1,4 @@
+- bump: patch
+  changes:
+    changed:
+      - Moved computed income variables to their respective section to avoid showing them as inputs.
