+++ policyengine_us/variables/household/income/person/capital_gains/capital_gains.py
@@ -1,0 +1,14 @@
+from policyengine_us.model_api import *
+
+
+class capital_gains(Variable):
+    value_type = float
+    entity = Person
+    label = "capital gains"
+    unit = USD
+    documentation = "Net gain from disposition of property."
+    definition_period = YEAR
+    adds = [
+        "short_term_capital_gains",
+        "long_term_capital_gains",
+    ]
