+++ policyengine_us/variables/household/income/person/pensions/pension_income.py
@@ -1,0 +1,14 @@
+from policyengine_us.model_api import *
+
+
+class pension_income(Variable):
+    value_type = float
+    entity = Person
+    label = "pension income"
+    unit = USD
+    documentation = "Income from pensions, annuitities, life insurance or endowment contracts."
+    definition_period = YEAR
+    adds = [
+        "tax_exempt_pension_income",
+        "taxable_pension_income",
+    ]
